+++ src/main/kotlin/omok/model/game/PlacementState.kt
@@ -1,4 +1,4 @@
-package omok.model
+package omok.model.game
 
 enum class PlacementState {
     WIN,
